refactor(models): relocate non-models out of src/models/common
Shuffle files that were sitting in models/common but weren't actually
model modules:

- character-storage → src/services/ (I/O, not a catalog)
- icons + GOLD_ICON → src/lib/ (low-level primitive; GOLD_ICON also
  moves off gear/equipment)
- character.ts + character-stats.ts → src/character/ (Character domain
  has its own folder now, pairing the type with its derived stats)
- compute fns extracted from abilities.ts + skills.ts → src/character/
  modifiers.ts (computeModifier/formatModifier/computeProficiencyBonus/
  computeSkillModifier — shared d20 math, not per-entity)
- damage-types.ts → src/models/common/damage/{damage-types,index}.ts
  (damage is shared by weapons + spells; subfolder leaves room to grow)
- gear/* → src/models/common/gear/* (regroup under common, which was
  always the right home)

Also inline recommendedScores into class-details.json and expose via
classes.get({id}).recommendedScores; promote standardArray to its own
JSON in src/data/class/. Drops the redundant recommended-scores module.

No API changes — this is pure relocation. Models whose files moved keep
the same exported object/type shape. All 398 existing tests still pass.
