+++ src/character/character.ts
@@ -1,10 +1,10 @@
 import type { CharacterResource } from "src/models/class/class-resources";
 import type { CharacterClass } from "src/models/class/classes";
-import type { Equipment } from "src/models/gear/equipment";
+import type { AbilityScores } from "src/models/common/abilities";
+import type { Equipment } from "src/models/common/gear/equipment";
 import type { Background } from "src/models/origin/backgrounds";
 import type { Species } from "src/models/origin/species";
 import type { Spell } from "src/models/spells/spells";
-import type { AbilityScores } from "./abilities";
 
 export type Character = {
   id: string;
