+++ src/components/character-sheet/character-sheet.tsx
@@ -1,18 +1,18 @@
 import c from "classnames";
 import { ScreenFlash, Stepper } from "elements";
 import { useState } from "react";
+import {
+  computeArmorClass,
+  computeInitiative,
+  computeSpellAttack,
+} from "src/character/character-stats";
 import { Inventory } from "src/components/inventory/inventory";
 import { ProficiencyGrid } from "src/components/proficiency-grid/proficiency-grid";
 import { useScrollDirection } from "src/hooks/use-scroll-direction";
 import { applyRest } from "src/models/class/class-resources";
 import { classes } from "src/models/class/classes";
-import {
-  computeArmorClass,
-  computeInitiative,
-  computeSpellAttack,
-} from "src/models/common/character-stats";
-import { saveCharacter } from "src/models/common/character-storage";
 import { spells } from "src/models/spells/spells";
+import { saveCharacter } from "src/services/character-storage";
 import { ActionBar } from "./action-bar";
 import { CharacterOverview } from "./character-overview";
 import styles from "./character-sheet.module.css";
@@ -24,8 +24,8 @@ import { SpellBook } from "./spell-book";
 import { SpellCards } from "./spell-cards";
 import { WeaponMastery } from "./weapon-mastery";
 
+import type { Character } from "src/character/character";
 import type { RestType } from "src/models/class/class-resources";
-import type { Character } from "src/models/common/character";
 
 const STEP_LABELS = ["Stats", "Combat", "Explore", "Spells & Skills", "Gear"];
