+++ src/components/character-sheet/spell-book/spell-book.tsx
@@ -5,7 +5,7 @@ import { SpellCards } from "src/components/character-sheet/spell-cards/spell-car
 import descriptionPopoverStyles from "src/elements/style/description-popover.module.css";
 import { useArrowOffset } from "src/hooks/use-arrow-offset";
 import { useExpandable } from "src/hooks/use-expandable";
-import { resolveIconPath } from "src/models/common/icons";
+import { resolveIconPath } from "src/lib/icons";
 import styles from "./spell-book.module.css";
 
 import type { Spell } from "src/models/spells/spells";
