--- src/data/class/recommended-scores.json
@@ -1,87 +1,0 @@
-{
-  "standardArray": [15, 14, 13, 12, 10, 8],
-  "recommended": {
-    "barbarian": {
-      "str": 15,
-      "dex": 13,
-      "con": 14,
-      "int": 8,
-      "wis": 12,
-      "cha": 10
-    },
-    "bard": { "str": 8, "dex": 14, "con": 13, "int": 10, "wis": 12, "cha": 15 },
-    "cleric": {
-      "str": 13,
-      "dex": 10,
-      "con": 14,
-      "int": 8,
-      "wis": 15,
-      "cha": 12
-    },
-    "druid": {
-      "str": 8,
-      "dex": 13,
-      "con": 14,
-      "int": 12,
-      "wis": 15,
-      "cha": 10
-    },
-    "fighter": {
-      "str": 15,
-      "dex": 13,
-      "con": 14,
-      "int": 8,
-      "wis": 12,
-      "cha": 10
-    },
-    "monk": { "str": 12, "dex": 15, "con": 13, "int": 8, "wis": 14, "cha": 10 },
-    "paladin": {
-      "str": 15,
-      "dex": 10,
-      "con": 13,
-      "int": 8,
-      "wis": 12,
-      "cha": 14
-    },
-    "ranger": {
-      "str": 12,
-      "dex": 15,
-      "con": 13,
-      "int": 8,
-      "wis": 14,
-      "cha": 10
-    },
-    "rogue": {
-      "str": 8,
-      "dex": 15,
-      "con": 14,
-      "int": 12,
-      "wis": 10,
-      "cha": 13
-    },
-    "sorcerer": {
-      "str": 8,
-      "dex": 13,
-      "con": 14,
-      "int": 10,
-      "wis": 12,
-      "cha": 15
-    },
-    "warlock": {
-      "str": 8,
-      "dex": 13,
-      "con": 14,
-      "int": 10,
-      "wis": 12,
-      "cha": 15
-    },
-    "wizard": {
-      "str": 8,
-      "dex": 13,
-      "con": 14,
-      "int": 15,
-      "wis": 12,
-      "cha": 10
-    }
-  }
-}
