+++ src/data/class/standard-array.json
@@ -1,0 +1,1 @@
+[15, 14, 13, 12, 10, 8]
