+++ src/lib/icons.ts
@@ -1,5 +1,7 @@
 export type IconVariant = "BLACK" | "WHITE";
 
+export const GOLD_ICON = "vol1/icon-vol1_63";
+
 export function resolveIconPath(
   pseudoPath: string,
   { variant = "BLACK" }: { variant?: IconVariant } = {},
