+++ src/models/common/abilities.ts
@@ -28,15 +28,3 @@ export const abilities = {
     return DATA;
   },
 };
-
-export function computeProficiencyBonus(level: number): number {
-  return Math.ceil(level / 4) + 1;
-}
-
-export function computeModifier(score: number): number {
-  return Math.floor((score - 10) / 2);
-}
-
-export function formatModifier(mod: number): string {
-  return mod >= 0 ? `+${mod}` : `${mod}`;
-}
