+++ src/models/common/damage/damage-types.ts
