+++ src/models/common/damage/index.ts
@@ -1,0 +1,1 @@
+export type { DamageType } from "./damage-types";
