+++ src/models/common/gear/armor.ts
