+++ src/models/common/gear/starting-equipment.test.ts
