+++ src/models/common/gear/weapons.test.ts
