+++ src/models/common/skills.test.ts
@@ -1,4 +1,4 @@
-import { computeSkillModifier, skills } from "./skills";
+import { skills } from "./skills";
 
 import type { AbilityName } from "./abilities";
 
@@ -61,44 +61,5 @@ describe("skills.findAll()", () => {
       const result = skills.findAll({ ability: ability as AbilityName });
       expect(result.map((s) => s.name)).toEqual(expected);
     });
-  });
-});
-
-describe("computeSkillModifier()", () => {
-  it("returns ability modifier when not proficient", () => {
-    expect(
-      computeSkillModifier({
-        abilityModifier: 2,
-        proficiencyBonus: 3,
-        isProficient: false,
-      }),
-    ).toBe(2);
-  });
-
-  it("returns ability modifier + proficiency bonus when proficient", () => {
-    expect(
-      computeSkillModifier({
-        abilityModifier: 2,
-        proficiencyBonus: 3,
-        isProficient: true,
-      }),
-    ).toBe(5);
-  });
-
-  it("handles negative ability modifiers", () => {
-    expect(
-      computeSkillModifier({
-        abilityModifier: -1,
-        proficiencyBonus: 3,
-        isProficient: true,
-      }),
-    ).toBe(2);
-    expect(
-      computeSkillModifier({
-        abilityModifier: -1,
-        proficiencyBonus: 3,
-        isProficient: false,
-      }),
-    ).toBe(-1);
   });
 });
